requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.3
crispy-bootstrap5==0.7
Django==4.2
django-celery-results==2.5.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.6
